==> Final-Project/1-Party_programs.ipynb <==
import pandas as pd
from bs4 import BeautifulSoup
import requests
import pandas as pd
import re
from selenium import webdriver
from time import sleep
import numpy as np
import src.cleaning as cl
import src.scraping as sc

### Unidas Podemos 

#### Scraping

scraping = {
    "https://podemos.info/bloque/horizonte-verde-nuevo-modelo-industrial/": "Horizonte Verde y Nuevo Modelo Industrial",
    "https://podemos.info/bloque/garantias-democraticas-ciudadania/": "Garantías Democráticas y de Ciudadanía",
    "https://podemos.info/bloque/garantias-justicia-fiscal/": "Garantías de Justicia Fiscal",
    "https://podemos.info/bloque/horizonte-morado-economia-los-cuidados/": "Horizonte Morado y Economía de los Cuidados",
    "https://podemos.info/bloque/garantias-justicia-laboral-pensiones/": "Garantías de Justicia Laboral y Pensiones",
    "https://podemos.info/bloque/garantias-justicia-territorial/": "Garantías de Justicia Territorial",
    "https://podemos.info/bloque/horizonte-digital-nueva-economia/": "Horizonte Digital y Nueva Economía",
    "https://podemos.info/bloque/garantias-justicia-social/": "Garantías de Justicia Social"
}

podemos_total = []
for key, value in scraping.items():
    result = sc.scraping_podemos(key, value)
    podemos_total.append(result)

unidas_podemos_total = podemos_total[0] + podemos_total[1] + podemos_total[2] + podemos_total[3] + podemos_total[4] + podemos_total[5] + podemos_total[6] + podemos_total[7]

df_unidas_podemos = pd.DataFrame(unidas_podemos_total)

df_unidas_podemos.reset_index(inplace = True)

df_unidas_podemos.rename(columns = {"index" : "id", 0 : "phrases"}, inplace = True)

to_delete = ("Email: contacto@podemos.info", "C/ Francisco Villaespesa, 1828017 Madrid", 
             "Comparte esta medida", "(Véase el apartado «Garantías de Justicia Territorial»).m", "podemos.info")

for e in to_delete:
    df_unidas_podemos = df_unidas_podemos[df_unidas_podemos["phrases"].str.contains(e)==False]

df_unidas_podemos.to_csv("./data/programs/csv-s/unidas_podemos.csv", index = False)

### Mas Madrid

df_mas_madrid = pd.read_csv("./data/programs/csv-s/mas_madrid.csv")

### PP

df_pp = pd.read_csv("./data/programs/csv-s/pp.csv")

### VOX

df_vox = pd.read_csv("./data/programs/csv-s/vox.csv")

df_vox

df_vox.drop(columns = ["Unnamed: 0"], inplace = True)

df_vox.reset_index(inplace = True)

df_vox

df_vox.rename(columns = {'index': 'id', 'VOX': 'phrases'}, inplace = True)

df_vox

df_vox.to_csv("./data/programs/csv-s/vox.csv")

### Ciudadanos

ciudadanos = sc.scrape_ciudadanos()

df_ciudadanos = pd.DataFrame(ciudadanos)

df_ciudadanos.rename(columns = {"index": "id", 0: "phrases" }, inplace = True)

df_ciudadanos['phrases'].replace("\n", " ", regex = True, inplace = True)

df_ciudadanos.reset_index(inplace = True)

df_ciudadanos

df_ciudadanos.to_csv("./data/programs/csv-s/ciudadanos.csv", index = False)

### PSOE

df_psoe = pd.read_excel("./data/programs/pdf-s_and_words/psoe.xlsx")

df_psoe.reset_index(inplace = True)

df_psoe.rename(columns = {"index": "id", 'PSOE': "phrases" }, inplace = True)

cl.replace(df_psoe, 'phrases', r"- ", r"")

df_psoe.to_csv("./data/programs/csv-s/psoe.csv", index = False)

### Put all the programs together to insert them to the database

df_mas_madrid['party_id'] = 2

df_vox['party_id'] = 6

df_unidas_podemos['party_id'] = 5

df_ciudadanos['party_id'] = 1

df_pp['party_id'] = 3

df_psoe['party_id'] = 4

all_programs = pd.concat([df_mas_madrid, df_vox, df_unidas_podemos, df_ciudadanos, df_pp, df_psoe])

all_programs.drop(columns = ["id","index"], inplace = True)

all_programs = cl.create_index(all_programs)

cl.replace(all_programs, 'phrases', r"- ", r"")

cl.replace(all_programs, 'phrases', r"[\"]", r"'")

cl.replace(all_programs, 'phrases', r"%", "por ciento")

cl.replace(all_programs, 'phrases', r"«", "'")

cl.replace(all_programs, 'phrases', r"»", "'")

column_names = ["id", "party_id", "phrases"] 
all_programs = all_programs.reindex(columns = column_names)

all_programs = all_programs.dropna(how = "any", axis = 0)

all_programs.isnull().sum()

all_programs.to_csv("./data/programs/csv-s/all_programs.csv", index = False)

all_programs = pd.read_csv("./data/programs/csv-s/all_programs.csv")

### Create the content of the tables and candidates tables

parties = {"id": [1, 2, 3, 4, 5, 6], "name": ['Ciudadanos', 'Más Madrid', 'PP', 'PSOE', 'Unidas Podemos', 'Vox']}
parties_table = pd.DataFrame(parties)

parties_table.to_csv("./data/programs/csv-s/parties_table.csv", index = False)

candidates = {"id": [1, 2, 3, 4, 5, 6], "name": ["Ángel", "Edmundo", "Isabel", "Mónica", "Pablo", "Rocío"],
             "surname": ["Gabilondo", "Bal", "Díaz Ayuso", "García", "Iglesias", "Monasterio"], 
             "party_id": [4, 1, 3, 2, 5, 6]
             }

candidates_table = pd.DataFrame(candidates)

candidates_table.to_csv("./data/programs/csv-s/candidates.csv", index = False)
