==> Final-Project/2-Verba_results.ipynb <==
from selenium import webdriver
from time import sleep
import pandas as pd
import numpy as np
import src.scraping as sc
import src.cleaning as cl

### Getting info from Verba

I decided to make two dictionaries, not to do all the searches at the same time and have problems with the website. 

# Dictionary to store the info for the searches
search_verba_parties = {
    "Unidas Podemos": "10/03/2021 - 04/05/2021",
    "Podemos": "10/03/2021 - 04/05/2021",
    "PP": "10/03/2021 - 04/05/2021",
    "Partido Popular": "10/03/2021 - 04/05/2021",
    "Más Madrid": "10/03/2021 - 04/05/2021",
    "PSOE": "10/03/2021 - 04/05/2021",
    "Partido Socialista": "10/03/2021 - 04/05/2021",
    "Vox": "10/03/2021 - 04/05/2021", 
    "Ciudadanos": "10/03/2021 - 04/05/2021"
}

search_verba_candidates = {
    "Pablo Iglesias": "10/03/2021 - 04/05/2021",
    "Iglesias": "10/03/2021 - 04/05/2021",
    "candidato de Unidas Podemos": "10/03/2021 - 04/05/2021",
    "Isabel Díaz Ayuso": "10/03/2021 - 04/05/2021",
    "Ayuso": "10/03/2021 - 04/05/2021",
    "candidata del Partido Popular": "10/03/2021 - 04/05/2021",
    "candidata del PP": "10/03/2021 - 04/05/2021",
    "Mónica García": "10/03/2021 - 04/05/2021",
    "García": "10/03/2021 - 04/05/2021",
    "candidata de Más Madrid": "10/03/2021 - 04/05/2021",
    "Ángel Gabilondo": "10/03/2021 - 04/05/2021",
    "Gabilondo": "10/03/2021 - 04/05/2021",
    "candidato del PSOE": "10/03/2021 - 04/05/2021",
    "Rocío Monasterio": "10/03/2021 - 04/05/2021",
    "Monasterio": "10/03/2021 - 04/05/2021",
    "candidata de Vox": "10/03/2021 - 04/05/2021",
    "Edmundo Bal": "10/03/2021 - 04/05/2021",
    "Bal": "10/03/2021 - 04/05/2021",
    "candidato de Ciudadanos": "10/03/2021 - 04/05/2021"
}

# Loop to do the searches with the info from the dictionary
# COMMENTED OUT NOT TO RUN IT AGAIN 
#for key, value in search_verba_parties.items():
    #sc.get_verba(key, value)

#for key, value in search_verba_candidates.items():
# COMMENTED OUT NOT TO RUN IT AGAIN 
    #sc.get_verba(key, value)

### Cleaning the dataframes

### Ángel Gabilondo

df_gabilondo1 = pd.read_csv("./data/verba_results/verba_candidates/angel_gabilondo1.csv")
df_gabilondo2 = pd.read_csv("./data/verba_results/verba_candidates/angel_gabilondo2.csv") 
df_gabilondo3 = pd.read_csv("./data/verba_results/verba_candidates/angel_gabilondo3.csv")

# Create a dataframe droping duplicates from the search
df_gabilondo = pd.concat([df_gabilondo1, df_gabilondo2, df_gabilondo3]).drop_duplicates().reset_index(drop = True)

df_verba_gabilondo = cl.clean_verba(df_gabilondo, 4, 1, 1)

### Edmundo Val

df_edmundo1 = pd.read_csv("./data/verba_results/verba_candidates/edmundo_bal1.csv")
df_edmundo2 = pd.read_csv("./data/verba_results/verba_candidates/edmundo_bal2.csv")
df_edmundo3 = pd.read_csv("./data/verba_results/verba_candidates/edmundo_bal3.csv")

# Create a dataframe droping duplicates from the search
df_edmundo = pd.concat([df_edmundo1, df_edmundo2, df_edmundo3]).drop_duplicates().reset_index(drop = True)

df_verba_edmundo = cl.clean_verba(df_edmundo, 1, 2, 1)

### Isabel Díaz Ayuso

df_ayuso1 = pd.read_csv("./data/verba_results/verba_candidates/isabel_diaz_ayuso1.csv")
df_ayuso2 = pd.read_csv("./data/verba_results/verba_candidates/isabel_diaz_ayuso2.csv")
df_ayuso3 = pd.read_csv("./data/verba_results/verba_candidates/isabel_diaz_ayuso3.csv")
df_ayuso4 = pd.read_csv("./data/verba_results/verba_candidates/isabel_diaz_ayuso4.csv")

df_ayuso = pd.concat([df_ayuso1, df_ayuso2, df_ayuso3, df_ayuso4]).drop_duplicates().reset_index(drop = True)

df_verba_ayuso = cl.clean_verba(df_ayuso, 3, 3, 1)

### Mónica García 

df_garcia1 = pd.read_csv("./data/verba_results/verba_candidates/monica_garcia1.csv")
df_garcia2 = pd.read_csv("./data/verba_results/verba_candidates/monica_garcia2.csv")
df_garcia3 = pd.read_csv("./data/verba_results/verba_candidates/monica_garcia3.csv")

df_garcia = pd.concat([df_garcia1, df_garcia2, df_garcia3]).drop_duplicates().reset_index(drop = True)

df_verba_garcia = cl.clean_verba(df_garcia, 2, 4, 1)

df_verba_garcia = df_verba_garcia[(df_verba_garcia['phrase'].str.contains("Mónica García")) | (df_verba_garcia['phrase'].str.contains("Más Madrid"))]

df_verba_garcia.drop(columns = 'id', inplace = True)

cl.create_index(df_verba_garcia)

### Pablo Iglesias

df_iglesias1 = pd.read_csv("./data/verba_results/verba_candidates/pablo_iglesias1.csv")
df_iglesias2 = pd.read_csv("./data/verba_results/verba_candidates/pablo_iglesias2.csv")
df_iglesias3 = pd.read_csv("./data/verba_results/verba_candidates/pablo_iglesias3.csv")

df_iglesias = pd.concat([df_iglesias1, df_iglesias2, df_iglesias3]).drop_duplicates().reset_index(drop = True)

df_verba_iglesias = cl.clean_verba(df_iglesias, 5, 5, 1)

### Rocío Monasterio

df_monasterio1 = pd.read_csv("./data/verba_results/verba_candidates/rocio_monasterio1.csv")
df_monasterio2 = pd.read_csv("./data/verba_results/verba_candidates/rocio_monasterio2.csv")
df_monasterio3 = pd.read_csv("./data/verba_results/verba_candidates/rocio_monasterio3.csv")

df_monasterio = pd.concat([df_monasterio1, df_monasterio2, df_monasterio3]).drop_duplicates().reset_index(drop = True)

df_verba_monasterio = cl.clean_verba(df_monasterio, 6, 6, 1)

# I will discard all the entries where monasterio appears in lower case
df_verba_monasterio = df_verba_monasterio[~df_verba_monasterio['phrase'].str.contains('monasterio')]

### Ciudadanos

df_ciudadanos = pd.read_csv("./data/verba_results/verba_parties/verba_ciudadanos.csv")
df_verba_ciudadanos = cl.clean_verba(df_ciudadanos, 1, 2, 0)

# I will only keep the cases in which Podemos appears in upper cases, as in lower cases it should be the plural of citizen
df_verba_ciudadanos = df_verba_ciudadanos[~df_verba_ciudadanos['phrase'].str.contains('ciudadanos')]

### Más Madrid

df_mas_madrid = pd.read_csv("./data/verba_results/verba_parties/verba_mas_madrid.csv")
df_verba_mas_madrid = cl.clean_verba(df_mas_madrid, 2, 4, 0)

### PP

df_pp1 = pd.read_csv("./data/verba_results/verba_parties/verba_partido_popular.csv")
df_pp2 = pd.read_csv("./data/verba_results/verba_parties/verba_PP.csv")

df_pp = pd.concat([df_pp1, df_pp2]).drop_duplicates().reset_index(drop = True)

df_verba_pp = cl.clean_verba(df_pp, 3, 3, 0)

### PSOE

df_psoe1 = pd.read_csv("./data/verba_results/verba_parties/verba_partido_socialista.csv")
df_psoe2 = pd.read_csv("./data/verba_results/verba_parties/verba_PSOE.csv")

df_psoe = pd.concat([df_psoe1, df_psoe2]).drop_duplicates().reset_index(drop = True)

df_verba_psoe = cl.clean_verba(df_psoe, 4, 1, 0)

### Unidas Podemos

df_unidas_podemos1 = pd.read_csv("./data/verba_results/verba_parties/verba_unidas_podemos1.csv")
df_unidas_podemos2 = pd.read_csv("./data/verba_results/verba_parties/verba_unidas_podemos2.csv")

df_unidas_podemos = pd.concat([df_unidas_podemos1, df_unidas_podemos2]).drop_duplicates().reset_index(drop = True)

df_verba_unidas_podemos = cl.clean_verba(df_unidas_podemos, 5, 5, 0)

# I will only keep the cases in which Podemos appears in upper cases, as in lower cases it should be the verb
df_verba_unidas_podemos = df_verba_unidas_podemos[~df_verba_unidas_podemos['phrase'].str.contains("podemos")]

### Vox

df_vox = pd.read_csv("./data/verba_results/verba_parties/verba_vox.csv")
df_verba_vox = cl.clean_verba(df_vox, 6, 6, 0)

### Concat all the dataframes to insert them into the database

all_verba_results = pd.concat([df_verba_gabilondo, df_verba_edmundo, df_verba_ayuso, df_verba_garcia, df_verba_iglesias, df_verba_monasterio, df_verba_ciudadanos, df_verba_mas_madrid, df_verba_pp, df_verba_psoe, df_verba_unidas_podemos, df_verba_vox])

all_verba_results.drop(columns = ["id", "index"], inplace = True)

all_verba_results = cl.create_index(all_verba_results)

cl.replace(all_verba_results, 'phrase', r"[\"]", r"'")

cl.replace(all_verba_results, 'phrase', r"%", "por ciento")

cl.replace(all_verba_results, 'phrase', r"%", "por ciento")

all_verba_results.info()

all_verba_results.to_csv("./data/verba_results/verba_candidates/all_verba_results.csv")
