==> Final-Project/3-SQL_database.ipynb <==
import pandas as pd
import sqlalchemy as alch
import config.configuration as con
from getpass import getpass
import os
import dotenv
dotenv.load_dotenv()

parties = pd.read_csv("./data/programs/csv-s/parties_table.csv")

candidates = pd.read_csv("./data/programs/csv-s/candidates.csv")

all_programs = pd.read_csv("./data/programs/csv-s/all_programs.csv")

verba = pd.read_csv("./data/verba_results/verba_candidates/all_verba_results.csv")

verba.drop(columns = "Unnamed: 0", inplace = True)

### Create database and insert the data

password = getpass("Introduce tu pass de sql: ")
dbName="elections"
connectionData=f"mysql+pymysql://root:{password}@localhost/{dbName}"

engine = alch.create_engine(connectionData)
print("Conected")

con.engine.execute("""
    USE elections;
""")

engine.execute("""
    DROP TABLE IF EXISTS parties;
""")

engine.execute("""
    CREATE TABLE parties(
    id INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
    name VARCHAR(40) NOT NULL UNIQUE);
""")

for n, line in parties.iterrows():
    engine.execute(
        f"""
        INSERT INTO parties VALUES
        ({line['id']},"{line['name']}");
       """
    )

engine.execute("""
    DROP TABLE IF EXISTS candidates;
""")

engine.execute("""
    CREATE TABLE candidates(
    id INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
    name VARCHAR(40) NOT NULL,
    surname VARCHAR(40) NOT NULL,
    party_id INT NOT NULL,
    FOREIGN KEY(party_id) REFERENCES parties(id));
""")

for n, line in candidates.iterrows():
    engine.execute(
        f"""
        INSERT INTO candidates VALUES
        ({line['id']},"{line['name']}", "{line['surname']}",{line['party_id']});
       """
    )

engine.execute("""
    DROP TABLE IF EXISTS programs;
""")

engine.execute("""
    CREATE TABLE programs(
    id INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
    party_id INT NOT NULL,
    phrases VARCHAR(10000) NOT NULL,
    FOREIGN KEY(party_id) REFERENCES parties(id))
    ENGINE = INNODB;
""")

for n, line in all_programs.iterrows():
    engine.execute(
        f"""
        INSERT INTO programs VALUES
        ({line['id']}, {line['party_id']}, "{line['phrases']}");
       """
    )

engine.execute("""
    DROP TABLE IF EXISTS verba;
""")

engine.execute("""
    CREATE TABLE verba(
    id INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
    p0_c1 INT NOT NULL,
    party_id INT NOT NULL,
    candidate_id INT NOT NULL,
    phrase VARCHAR(10000) NOT NULL,
    date DATE NOT NULL,
    FOREIGN KEY(candidate_id) REFERENCES candidates(id), 
    FOREIGN KEY(party_id) REFERENCES parties(id))
    ENGINE = INNODB;   
""")

for n, line in verba.iterrows():
    engine.execute(
        f"""
        INSERT INTO verba VALUES
        ({line['id']},{line['p0_c1']}, {line['party_id']}, {line['candidate_id']},"{line['phrase']}", "{line['date']}");
       """
    )
