==> Final-Project/4-Text_analysis.ipynb <==
import pandas as pd
import nltk
from nltk.tokenize import RegexpTokenizer
from nltk.corpus import stopwords
from nltk import FreqDist
from nltk.sentiment.vader import SentimentIntensityAnalyzer 
import spacy
from collections import Counter
from textblob import TextBlob
import sqlalchemy as alch
from getpass import getpass
from config.configuration import engine
import src.cleaning as cl
import src.get_data as gd
import dotenv
dotenv.load_dotenv()

password = getpass("Introduce tu pass de sql: ")
dbName="elections"
connectionData=f"mysql+pymysql://root:{password}@localhost/{dbName}"

engine = alch.create_engine(connectionData)
print("Conected")

### PROGRAMS

all_programs = gd.get_phrases()

verba = gd.verba_phrases()

### Tokenize phrases

nltk.download('punkt')

all_programs.dropna(axis = 0, how = 'any', inplace = True) 

all_programs['tokenized'] = all_programs['phrases'].apply(cl.tokenize)

### Apply stop_words

nltk.download('stopwords')

all_programs['tokenized'] = all_programs['tokenized'].apply(cl.stop_words)

all_programs

#spanish_stemmer = SnowballStemmer('spanish')

#spanish_stemmer.stem("libertad, libre, libres")

#spanish_stemmer.stem("libertad, libre, libres, libertad, libre")

Between Stemming and Lematization, the latter works better. However, unfortunately there is no option to do it in Spanish. 

Although there is the option of translating the text, in this case I consider it important to keep it in the original language. 

I tried the SnoballStemmer, which is available in Spanish, but as it can be seen above the results are very good. It was supposed to take the root of the word and categorize all as the same. Instead, it puts them all as the same and deletes a part of the last one. 

Therefore, I will perform the search of the most common words in the column where I applied the tokenization and the stop words.

### Count words using value_counts

# Create a dataframe with the word count of all the programs together
df_word_count = cl.count_values(all_programs, 'tokenized')

# Make a list with the words to delete (The ones that would appear in the first 30)
to_delete = ["La", "Se"]

# Apply the function to delete them
df_word_count = cl.delete_extra_words(df_word_count, to_delete)

df_word_count = cl.for_streamlit(df_word_count)

### Using the Spacy library

nlp = spacy.load('es_core_news_sm')

all_common_verbs = cl.spacy_verb(all_programs, 'tokenized')

all_common_nouns = cl.spacy_noun(all_programs, 'tokenized')

all_programs[all_programs['phrases'].str.contains("persona")]

The results of the search made using value_counts and Spacy are different. I would say this happens due to errors in the entity recognition of spacy, which doesn't always correctly put the tag (noun, verb...) to the word. 

I will use both methods (the one with value_counts() and the spacy library to learn how to use it. 

### Analyze each party

#### PP

pp_phrases = gd.phrases("PP")

# Create a column to apply tokenization
pp_phrases['tokenized'] = pp_phrases['phrases'].apply(cl.tokenize)

# Use the stop words in the column 
pp_phrases['tokenized'] = pp_phrases['tokenized'].apply(cl.stop_words)

# Apply the first function that creates a dataframe with words capitalized and the frequency
pp_phrases_count = cl.count_values(pp_phrases, 'tokenized')

# Make a list with the words to delete
to_delete = ["Años", "Ello", "En", "La", "Puedan", "Través", "Especialmente", 
             "Nuevas", "Vamos", "Además", "Especialmente", "Una", "Aquellos", "Cada"]

# Apply function to delete words
pp_phrases_count = cl.delete_extra_words(pp_phrases_count, to_delete)

# Prepare the dataframe to make a graph in streamlit
pp_phrases_count = cl.for_streamlit(pp_phrases_count)

pp_phrases_count.to_csv("./data/programs/csv-s/pp_phrases_count.csv")

pp_verbs = cl.spacy_verb(pp_phrases, 'tokenized')

#pp_verbs.to_csv("./data/programs/csv-s/pp_verbs.csv")

pp_nouns = cl.spacy_noun(pp_phrases, 'tokenized')

#pp_nouns.to_csv("./data/programs/csv-s/pp_nouns.csv")

#### PSOE

psoe_phrases = gd.phrases("PSOE")

psoe_phrases['tokenized'] = psoe_phrases['phrases'].apply(cl.tokenize)

psoe_phrases['tokenized'] = psoe_phrases['tokenized'].apply(cl.stop_words)

psoe_phrases_count = cl.count_values(psoe_phrases, 'tokenized')

to_delete = ["N", "La", "Así"]

psoe_phrases_count = cl.delete_extra_words(psoe_phrases_count, to_delete)

psoe_phrases_count = cl.for_streamlit(psoe_phrases_count)

#psoe_phrases_count.to_csv("./data/programs/csv-s/psoe_phrases_count.csv")

psoe_verbs = cl.spacy_verb(psoe_phrases, 'tokenized')

#psoe_verbs.to_csv("./data/programs/csv-s/psoe_verbs.csv")

psoe_nouns = cl.spacy_noun(psoe_phrases, 'tokenized')

#psoe_nouns.to_csv("./data/programs/csv-s/psoe_nouns.csv")

#### Más Madrid

mas_madrid_phrases = gd.phrases("Más Madrid")

mas_madrid_phrases['tokenized'] = mas_madrid_phrases['phrases'].apply(cl.tokenize)

mas_madrid_phrases['tokenized'] = mas_madrid_phrases['tokenized'].apply(cl.stop_words)

mas_madrid_phrases_count = cl.count_values(psoe_phrases, 'tokenized')

to_delete = ["Ción", "N", "La", "Dad", "Así", "Dos", "Co"]

mas_madrid_phrases_count = cl.delete_extra_words(mas_madrid_phrases_count, to_delete)

mas_madrid_phrases_count = cl.for_streamlit(mas_madrid_phrases_count)

#mas_madrid_phrases_count.to_csv("./data/programs/csv-s/mas_madrid_phrases_count.csv")

mas_madrid_verbs = cl.spacy_verb(mas_madrid_phrases, 'tokenized')

#mas_madrid_verbs.to_csv("./data/programs/csv-s/mas_madrid_verbs.csv")

mas_madrid_nouns = cl.spacy_noun(mas_madrid_phrases, 'tokenized')

#mas_madrid_nouns.to_csv("./data/programs/csv-s/mas_madrid_nouns.csv")

#### Unidas Podemos

unidas_podemos_phrases = gd.phrases("Unidas Podemos")

unidas_podemos_phrases['tokenized'] = unidas_podemos_phrases['phrases'].apply(cl.tokenize)

unidas_podemos_phrases['tokenized'] = unidas_podemos_phrases['tokenized'].apply(cl.stop_words)

unidas_podemos_phrases_count = cl.count_values(unidas_podemos_phrases, 'tokenized')

to_delete = ["Cada", "Ciento", "Manera", "Menos", "Ello", "La", "Se", "Así", "Con", "Para", "En", "Por", "Puedan", "Todas", "Fin", "Si", "Asimismo", "000"]

unidas_podemos_phrases_count = cl.delete_extra_words(unidas_podemos_phrases_count, to_delete)

unidas_podemos_phrases_count = cl.for_streamlit(unidas_podemos_phrases_count)

#unidas_podemos_phrases_count.to_csv("./data/programs/csv-s/unidas_podemos_phrases_count.csv")

unidas_podemos_verbs = cl.spacy_verb(unidas_podemos_phrases, 'tokenized')

#unidas_podemos_verbs.to_csv("./data/programs/csv-s/unidas_podemos_verbs.csv")

unidas_podemos_nouns = cl.spacy_noun(unidas_podemos_phrases, 'tokenized')

#unidas_podemos_nouns.to_csv("./data/programs/csv-s/unidas_podemos_nouns.csv")

#### Ciudadanos

ciudadanos_phrases = gd.phrases("Ciudadanos")

ciudadanos_phrases['tokenized'] = ciudadanos_phrases['phrases'].apply(cl.tokenize)

ciudadanos_phrases['tokenized'] = ciudadanos_phrases['tokenized'].apply(cl.stop_words)

ciudadanos_phrases_count = cl.count_values(ciudadanos_phrases, 'tokenized')

to_delete = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "Los", "Impulsaremos", "Garantizaremos", "Apostaremos",
            "Través", "Puedan", "Aseguraremos"]

ciudadanos_phrases_count = cl.delete_extra_words(ciudadanos_phrases_count, to_delete)

ciudadanos_phrases_count = cl.for_streamlit(ciudadanos_phrases_count)

#ciudadanos_phrases_count.to_csv("./data/programs/csv-s/ciudadanos_phrases_count.csv")

ciudadanos_verbs = cl.spacy_verb(ciudadanos_phrases, 'tokenized')

#ciudadanos_verbs.to_csv("./data/programs/csv-s/ciudadanos_verbs.csv")

ciudadanos_nouns = cl.spacy_noun(ciudadanos_phrases, 'tokenized')

#ciudadanos_nouns.to_csv("./data/programs/csv-s/ciudadanos_nouns.csv")

#### Vox

vox_phrases = gd.phrases("Vox")

vox_phrases['tokenized'] = vox_phrases['phrases'].apply(cl.tokenize)

vox_phrases['tokenized'] = vox_phrases['tokenized'].apply(cl.stop_words)

vox_phrases_count = cl.count_values(vox_phrases, 'tokenized')

to_delete = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "Deben", "Queda",
             "Frente", "Especialmente", "Cometan", "Su", "Pequeñas", "Adecuado"]

vox_phrases_count = cl.delete_extra_words(vox_phrases_count, to_delete)

vox_phrases_count = cl.for_streamlit(vox_phrases_count)

#vox_phrases_count.to_csv("./data/programs/csv-s/vox_phrases_count.csv")

vox_verbs = cl.spacy_verb(vox_phrases, 'tokenized')

#vox_verbs.to_csv("./data/programs/csv-s/vox_verbs.csv")

vox_nouns = cl.spacy_noun(vox_phrases, 'tokenized')

#vox_nouns.to_csv("./data/programs/csv-s/vox_nouns.csv")

### VERBA RESULTS

### Ángel Gabilondo

verba_gabilondo = gd.verba_candidate(1)

verba_gabilondo['tokenized'] = verba_gabilondo['phrase'].apply(cl.tokenize)

verba_gabilondo['tokenized'] = verba_gabilondo['tokenized'].apply(cl.stop_words)

verba_gabilondo_count = cl.count_values(verba_gabilondo, 'tokenized')

to_delete = ["El", "Si", "Y", "La", "Dice", "Hoy", "Presentado", "En", "Va", "Hora"]

verba_gabilondo_count = cl.delete_extra_words(verba_gabilondo_count, to_delete)

verba_gabilondo_count = cl.for_streamlit(verba_gabilondo_count)

#verba_gabilondo_count.to_csv("./data/verba_results/verba_candidates/gabilondo_verba.csv")

### Edmundo Bal 

verba_bal = gd.verba_candidate(2)

verba_bal['tokenized'] = verba_bal['phrase'].apply(cl.tokenize)

verba_bal['tokenized'] = verba_bal['tokenized'].apply(cl.stop_words)

verba_bal = cl.count_values(verba_bal, 'tokenized')

to_delete = ["El", "Y", "La", "Dice", "Tres", "Hoy", "En", "Si", "Tras", "Se", "Mañana"]

verba_bal_count = cl.delete_extra_words(verba_bal, to_delete)

verba_bal_count = cl.for_streamlit(verba_bal_count)

#verba_bal_count.to_csv("./data/verba_results/verba_candidates/bal_verba.csv")

### Isabel Díaz Ayuso

verba_ayuso = gd.verba_candidate(3)

verba_ayuso['tokenized'] = verba_ayuso['phrase'].apply(cl.tokenize)

verba_ayuso['tokenized'] = verba_ayuso['tokenized'].apply(cl.stop_words)

verba_ayuso = cl.count_values(verba_ayuso, 'tokenized')

to_delete = ["El", "Y", "La", "Dice", "Dos", "Hoy", "Si", "En", "No", "Sido", "Los"]

verba_ayuso_count = cl.delete_extra_words(verba_ayuso, to_delete)

verba_ayuso_count = cl.for_streamlit(verba_ayuso_count)

#verba_ayuso_count.to_csv("./data/verba_results/verba_candidates/ayuso_verba.csv")

### Mónica García 

verba_garcia = gd.verba_candidate(4)

verba_garcia['tokenized'] = verba_garcia['phrase'].apply(cl.tokenize)

verba_garcia['tokenized'] = verba_garcia['tokenized'].apply(cl.stop_words)

verba_garcia = cl.count_values(verba_garcia, 'tokenized')

to_delete = ["La", "En", "Días", "El", "Y", "4", "Hoy", "Es", "Años", "Semana", "Ciento", "Real", "Por", "Si", "Solo", "A", "Dos", "Va", "Dice", "Ahora"]

verba_garcia_count = cl.delete_extra_words(verba_garcia, to_delete)

verba_garcia_count = cl.for_streamlit(verba_garcia_count)

#verba_garcia_count.to_csv("./data/verba_results/verba_candidates/garcia_verba.csv")

### Pablo Iglesias

verba_iglesias = gd.verba_candidate(5)

verba_iglesias['tokenized'] = verba_iglesias['phrase'].apply(cl.tokenize)

verba_iglesias['tokenized'] = verba_iglesias['phrase'].apply(cl.tokenize)

verba_iglesias['tokenized'] = verba_iglesias['tokenized'].apply(cl.stop_words)

verba_iglesias = cl.count_values(verba_iglesias, 'tokenized')

to_delete = ["El", "Y", "Más", "Hoy", "Después", "La", "En", "Va", "Dice", "Aunque", "Si", "Dos"]

verba_iglesias_count = cl.delete_extra_words(verba_iglesias, to_delete)

verba_iglesias_count = cl.for_streamlit(verba_iglesias_count)

#verba_iglesias_count.to_csv("./data/verba_results/verba_candidates/iglesias_verba.csv")

### Rocío Monasterio

verba_monasterio = gd.verba_candidate(6)

verba_monasterio['tokenized'] = verba_monasterio['phrase'].apply(cl.tokenize)

verba_monasterio['tokenized'] = verba_monasterio['tokenized'].apply(cl.stop_words)

verba_monasterio = cl.count_values(verba_monasterio, 'tokenized')

to_delete = ["El", "La", "Y", "Tres", "Más", "Dice", "Si", "Ser", "Quiere", "En", "Hoy", "Ciento", "Va", "A"]

verba_monasterio_count = cl.delete_extra_words(verba_monasterio, to_delete)

verba_monasterio_count = cl.for_streamlit(verba_monasterio_count)

#verba_monasterio_count.to_csv("./data/verba_results/verba_candidates/monasterio_verba.csv")
