==> Final-Project/6- Sentiment_analysis.ipynb <==
import pandas as pd
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from nltk.corpus import stopwords
from textblob import TextBlob
import src.cleaning as cl

pp_verbs = pd.read_csv("./data/programs/csv-s/pp_verbs.csv")
vox_verbs = pd.read_csv("./data/programs/csv-s/vox_verbs.csv")
psoe_verbs = pd.read_csv("./data/programs/csv-s/psoe_verbs.csv")
ciudadanos_verbs = pd.read_csv("./data/programs/csv-s/ciudadanos_verbs.csv")
mas_madrid_verbs = pd.read_csv("./data/programs/csv-s/mas_madrid_verbs.csv")
unidas_podemos_verbs = pd.read_csv("./data/programs/csv-s/unidas_podemos_verbs.csv")

#sia = SentimentIntensityAnalyzer()

pp_verbs['english'] = pp_verbs['Word'].apply(cl.translate_into_english)

pp_verbs['sentiment_analysis'] = pp_verbs['english'].apply(cl.sentiment_analysis)

vox_verbs['english'] = vox_verbs['Word'].apply(cl.translate_into_english)

vox_verbs['sentiment_analysis'] = vox_verbs['english'].apply(cl.sentiment_analysis)

vox_verbs['sentiment_analysis'].mean()

pp_verbs['sentiment_analysis'].mean()

df['sentiment'] = df['text'].apply(lambda tweet: TextBlob(tweet).sentiment)

pp_verbs['text_blob'] = pp_verbs['english'].apply(lambda x: TextBlob(x).sentiment)

pp_verbs

vox_verbs['text_blob'] = vox_verbs['english'].apply(lambda x: TextBlob(x).sentiment)

vox_verbs
